% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JACA.R
\docType{package}
\name{JACA-package}
\alias{JACA}
\alias{JACA-package}
\title{JACA: Joint Association and Classification Analysis}
\description{
This package implements "Joint association and classification analysis of multi-view data".
Multi-view data refers to matched sets of measurements on the same subjects,
and JACA is a joint statistical framework that performs association and classification analysis simultaneously.
}
\details{
This package implements three main methods that can:
\itemize{
 \item Fit the JACA model.
 \item Use cross-validation function to automatically select parameters.
 \item Classify observations using the supplied model.
}
}
\references{
Zhang, Y., & Gaynanova, I. (2018). Joint association and classification analysis of multi-view data. \emph{arXiv preprint arXiv:1811.08511}.
}
\author{
Yunfeng Zhang and Irina Gaynanova
}
